[BACK-1161] Wire reject button for approved items
- Renamed and updated reject modal & form previously used for prospects
to RejectItem* components so that they can be used for both prospects and
approved items.

- Wired up the reject modal to the Reject button on the Live Corpus page.

- In an unrelated change, updated sidebar new tab component to display
full dates instead of relative dates as unit tests fail otherwise
(see details in JIRA - 1207).
