+++ src/curated-corpus/components/ApprovedItemListCard/ApprovedItemListCard.tsx
@@ -25,6 +25,8 @@ interface ApprovedItemListCardProps {
    */
   item: ApprovedCuratedCorpusItem;
 
+  onReject: () => void;
+
   onSchedule: () => void;
 }
 
@@ -32,7 +34,7 @@ export const ApprovedItemListCard: React.FC<ApprovedItemListCardProps> = (
   props
 ): JSX.Element => {
   const classes = useStyles();
-  const { item, onSchedule } = props;
+  const { item, onReject, onSchedule } = props;
 
   return (
     <Card className={classes.root}>
@@ -104,7 +106,7 @@ export const ApprovedItemListCard: React.FC<ApprovedItemListCardProps> = (
         <Button buttonType="positive" variant="text" onClick={onSchedule}>
           Schedule
         </Button>
-        <Button buttonType="negative" variant="text">
+        <Button buttonType="negative" variant="text" onClick={onReject}>
           Reject
         </Button>
         <Button buttonType="positive" variant="text">
