+++ src/curated-corpus/components/NewTabGroupedList/NewTabGroupedList.tsx
@@ -42,18 +42,10 @@ export const NewTabGroupedList: React.FC<NewTabGroupedListProps> = (
     ? MiniNewTabScheduleCard
     : MiniNewTabScheduleCard;
 
-  // Determine how to show the date. For today and tomorrow, use relative terms
-  const tomorrow = DateTime.local().plus({ days: 1 });
-  const scheduledDateObj = DateTime.fromFormat(scheduledDate, 'yyyy-MM-dd');
-  let displayDate: string | null;
-  if (scheduledDateObj <= tomorrow) {
-    displayDate = scheduledDateObj.toRelativeCalendar();
-  } else {
-    displayDate = scheduledDateObj
-      .setLocale('en')
-      .toLocaleString(DateTime.DATE_FULL);
-  }
-  // For beyond the next couple of days, use a standard date
+  const displayDate = DateTime.fromFormat(scheduledDate, 'yyyy-MM-dd')
+    .setLocale('en')
+    .toLocaleString(DateTime.DATE_FULL);
+
   return (
     <>
       <Typography
