+++ src/curated-corpus/components/ProspectListCard/ProspectListCard.test.tsx
@@ -6,6 +6,7 @@ import { ProspectListCard } from './ProspectListCard';
 
 describe('The ProspectListCard component', () => {
   let prospect: Prospect;
+  const onReject = jest.fn();
 
   beforeEach(() => {
     prospect = {
@@ -26,7 +27,7 @@ describe('The ProspectListCard component', () => {
   it('shows basic prospect information', () => {
     render(
       <MemoryRouter>
-        <ProspectListCard prospect={prospect} />
+        <ProspectListCard prospect={prospect} onReject={onReject} />
       </MemoryRouter>
     );
 
@@ -47,7 +48,8 @@ describe('The ProspectListCard component', () => {
   it('shows language correctly', () => {
     render(
       <MemoryRouter>
-        <ProspectListCard prospect={prospect} />
+        {' '}
+        <ProspectListCard prospect={prospect} onReject={onReject} />
       </MemoryRouter>
     );
 
@@ -57,7 +59,7 @@ describe('The ProspectListCard component', () => {
   it('shows topic correctly', () => {
     render(
       <MemoryRouter>
-        <ProspectListCard prospect={prospect} />
+        <ProspectListCard prospect={prospect} onReject={onReject} />
       </MemoryRouter>
     );
 
@@ -67,7 +69,8 @@ describe('The ProspectListCard component', () => {
   it('should render prospect card with excerpt', () => {
     render(
       <MemoryRouter>
-        <ProspectListCard prospect={prospect} />
+        {' '}
+        <ProspectListCard prospect={prospect} onReject={onReject} />
       </MemoryRouter>
     );
 
@@ -77,7 +80,7 @@ describe('The ProspectListCard component', () => {
   it('should render curated item card with the action buttons', () => {
     render(
       <MemoryRouter>
-        <ProspectListCard prospect={prospect} />
+        <ProspectListCard prospect={prospect} onReject={onReject} />
       </MemoryRouter>
     );
