+++ src/curated-corpus/components/index.ts
@@ -7,7 +7,7 @@ export { NextPrevPagination } from './NextPrevPagination/NextPrevPagination';
 export { ProspectListCard } from './ProspectListCard/ProspectListCard';
 export { RejectedItemListCard } from './RejectedItemListCard/RejectedItemListCard';
 export { RejectedItemSearchForm } from './RejectedItemSearchForm/RejectedItemSearchForm';
-export { RejectProspectModal } from './RejectProspectModal/RejectProspectModal';
-export { RejectProspectForm } from './RejectProspectForm/RejectProspectForm';
+export { RejectItemModal } from './RejectItemModal/RejectItemModal';
+export { RejectItemForm } from './RejectItemForm/RejectItemForm';
 export { ScheduleItemForm } from './ScheduleItemForm/ScheduleItemForm';
 export { ScheduleItemModal } from './ScheduleItemModal/ScheduleItemModal';
